# [DEV-5839] TAS/PSC filter speedup