--- a/usaspending_api/references/v2/views/filter_tree/filter_tree.py
+++ b/usaspending_api/references/v2/views/filter_tree/filter_tree.py
@@ -54,10 +54,15 @@ class FilterTree(metaclass=ABCMeta):
         raw_children = self.raw_search(ancestor_array + [retval.id])
         temp_children = [
             self._linked_node_from_data(ancestor_array + [retval.id], elem, filter_string, child_layers - 1)
-            for elem in raw_children if self.matches_filter(Node(**elem, count=0, children=None, ancestors=None), filter_string)
+            for elem in raw_children
         ]
+
         if child_layers:
-            children = temp_children
+            if filter_string:
+                temp_children = [elem for elem in temp_children if self.matches_filter(elem, filter_string)]
+                children = temp_children
+            else:
+                children = temp_children
         else:
             children = None
 
@@ -91,7 +96,6 @@ class FilterTree(metaclass=ABCMeta):
         pass
 
     def matches_filter(self, node: Node, filter_string) -> bool:
-
         if (filter_string.lower() in node.id.lower()) or (filter_string.lower() in node.description.lower()):
             return True
         if node.children: