--- a/usaspending_api/references/v2/views/filter_tree/psc_filter_tree.py
+++ b/usaspending_api/references/v2/views/filter_tree/psc_filter_tree.py
@@ -1,6 +1,8 @@
 import re
 
+from django.db.models import Q
 from string import ascii_uppercase, digits
+
 from usaspending_api.references.models import PSC
 from usaspending_api.references.v2.views.filter_tree.filter_tree import UnlinkedNode, FilterTree
 
@@ -16,16 +18,15 @@ PSC_GROUPS = {
 
 
 class PSCFilterTree(FilterTree):
-    def raw_search(self, tiered_keys):
+    def raw_search(self, tiered_keys, filter_string=None):
         if not self._path_is_valid(tiered_keys):
             return []
-
         if len(tiered_keys) == 0:
             return self._toptier_search()
         elif len(tiered_keys) == 1:
-            return self._psc_from_group(tiered_keys[0])
+            return self._psc_from_group(tiered_keys[0], filter_string)
         else:
-            return self._psc_from_parent(tiered_keys[-1])
+            return self._psc_from_parent(tiered_keys[-1], filter_string)
 
     def _path_is_valid(self, path: list) -> bool:
         if len(path) > 1:
@@ -39,22 +40,21 @@ class PSCFilterTree(FilterTree):
     def _toptier_search(self):
         return PSC_GROUPS.keys()
 
-    def _psc_from_group(self, group):
+    def _psc_from_group(self, group, filter_string: str):
         # The default regex value will match nothing
-        return self._psc_from_regex(PSC_GROUPS.get(group, {}).get("pattern") or "(?!)")
+        filters = [Q(code__iregex=PSC_GROUPS.get(group, {}).get("pattern") or "(?!)")]
+        return [{"id": object.code, "description": object.description} for object in PSC.objects.filter(*filters)]
 
-    def _psc_from_regex(self, regex):
-        return [
-            {"id": object.code, "description": object.description} for object in PSC.objects.filter(code__iregex=regex)
-        ]  # normally very unsafe, but regexes are not being supplied by the user
-
-    def _psc_from_parent(self, parent):
+    def _psc_from_parent(self, parent, filter_string: str):
         # two out of three branches of the PSC tree "jump" over 3 character codes
         desired_len = len(parent) + 2 if len(parent) == 2 and parent[0] != "A" else len(parent) + 1
-        return [
-            {"id": object.code, "description": object.description}
-            for object in PSC.objects.filter(length=desired_len, code__startswith=parent)
+        filters = [
+            Q(length=desired_len),
+            Q(code__startswith=parent),
         ]
+        if filter_string and desired_len == 4:
+            filters.append(Q(Q(code__icontains=filter_string) | Q(description__icontains=filter_string)))
+        return [{"id": object.code, "description": object.description} for object in PSC.objects.filter(*filters)]
 
     def unlinked_node_from_data(self, ancestors: list, data) -> UnlinkedNode:
         if len(ancestors) == 0:  # A tier zero search is returning an agency dictionary