--- a/usaspending_api/references/v2/views/filter_tree/tas_filter_tree.py
+++ b/usaspending_api/references/v2/views/filter_tree/tas_filter_tree.py
@@ -2,17 +2,17 @@ from usaspending_api.common.helpers.business_logic_helpers import cfo_presentati
 from usaspending_api.accounts.models import TreasuryAppropriationAccount, FederalAccount
 from usaspending_api.references.v2.views.filter_tree.filter_tree import UnlinkedNode, FilterTree
 from usaspending_api.references.models import ToptierAgency
-from django.db.models import Exists, OuterRef
+from django.db.models import Exists, OuterRef, Q
 
 
 class TASFilterTree(FilterTree):
-    def raw_search(self, tiered_keys):
+    def raw_search(self, tiered_keys, filter_search):
         if len(tiered_keys) == 0:
             return self._toptier_search()
         if len(tiered_keys) == 1:
-            return self._fa_given_agency(tiered_keys[0])
+            return self._fa_given_agency(tiered_keys[0], filter_search)
         if len(tiered_keys) == 2:
-            return self._tas_given_fa(tiered_keys[0], tiered_keys[1])
+            return self._tas_given_fa(tiered_keys[0], tiered_keys[1], filter_search)
         return []
 
     def _toptier_search(self):
@@ -35,19 +35,29 @@ class TASFilterTree(FilterTree):
     def _dictionary_from_agency(self, agency):
         return {"toptier_code": agency["toptier_code"], "name": agency["name"], "abbreviation": agency["abbreviation"]}
 
-    def _fa_given_agency(self, agency):
+    def _fa_given_agency(self, agency, filter_string):
+        filters = [
+            Q(has_faba=True),
+            Q(parent_toptier_agency__toptier_code=agency)
+        ]
+        if filter_string:
+            filters.append(Q(Q(federal_account_code__icontains=filter_string) | Q(account_title__icontains=filter_string)))
         return FederalAccount.objects.annotate(
             has_faba=Exists(faba_with_file_D_data().filter(treasury_account__federal_account=OuterRef("pk")))
-        ).filter(has_faba=True, parent_toptier_agency__toptier_code=agency)
+        ).filter(*filters)
 
-    def _tas_given_fa(self, agency, fed_account):
+    def _tas_given_fa(self, agency, fed_account, filter_string):
+        filters = [
+            Q(has_faba=True),
+            Q(federal_account__federal_account_code=fed_account),
+            Q(federal_account__parent_toptier_agency__toptier_code=agency)
+        ]
+        # if filter_string:
+        #     filters.append(
+        #         Q(Q(tas_rendering_label__icontains=filter_string) | Q(federal_account__account_title__icontains=filter_string)))
         return TreasuryAppropriationAccount.objects.annotate(
             has_faba=Exists(faba_with_file_D_data().filter(treasury_account=OuterRef("pk")))
-        ).filter(
-            has_faba=True,
-            federal_account__federal_account_code=fed_account,
-            federal_account__parent_toptier_agency__toptier_code=agency,
-        )
+        ).filter(*filters)
 
     def unlinked_node_from_data(self, ancestors: list, data) -> UnlinkedNode:
         if len(ancestors) == 0:  # A tier zero search is returning an agency dictionary