**Description:**
Modifies the behavior of the PSC reference endpoint, as well as performance improvements to both the TAS and PSC reference endpoints on the Advanced search page.

**Technical details:**
The main performance improvement is that instead of returning all the "children" of a given node from the DB and then filtering those children by the filter string, we add the filter string to the database call, which seems to have a pretty decent 
improvement to performance.
Performance Testing:

  | {{api_url}}/api/v2/references/filter_tree/tas/?depth=2&filter=073-1154 | {{api_url}}/api/v2/references/filter_tree/tas/?depth=2&filter=salaries | {{api_url}}/api/v2/references/filter_tree/tas/ | {{api_url}}/api/v2/references/filter_tree/tas/091/ | {{api_url}}/api/v2/references/filter_tree/tas/?depth=2&filter=stabilization | {{api_url}}/api/v2/references/filter_tree/tas/?depth=2&filter=institute
-- | -- | -- | -- | -- | -- | --
old   (sandbox) | 9.97 s | 6.37 s | 6.19 s | 1387 ms | 6.06 s | 7.07 s
new   (sandbox) | 818 ms | 1125 ms | 522 ms | 537 ms | 488 ms | 6.20 s


  | {{api_url}}/api/v2/references/filter_tree/psc/?depth=-1&filter=research | {{api_url}}/api/v2/references/filter_tree/psc/?depth=-1&filter=education | {{api_url}}/api/v2/references/filter_tree/psc/?depth=-1&filter=salaries | {{api_url}}/api/v2/references/filter_tree/psc/?depth=-1&filter=EQPT | {{api_url}}/api/v2/references/filter_tree/psc/ | {{api_url}}/api/v2/references/filter_tree/psc/?depth=-1&filter=engineer | {{api_url}}/api/v2/references/filter_tree/psc/?depth=-1&filter=agriculture
-- | -- | -- | -- | -- | -- | -- | --
new  (sandbox) | 801 ms | 494 ms | 877 ms | 1025 ms | 548 ms | 1307 ms | 1063 ms
old   (sandbox) | 7.40 s | 4.55 s | 4.76 s | 4.64 s | 4.50 s | 4.59 s | 4.50 s



**Requirements for PR merge:**

1. [X] Unit & integration tests updated
2. [X] API documentation updated
3. [ ] Necessary PR reviewers:
    - [ ] Backend
4. [X] Matview impact assessment completed
5. [X] Frontend impact assessment completed
6. [X] Data validation completed
7. [X] Appropriate Operations ticket(s) created
8. [X] Jira Ticket [DEV-5839](https://federal-spending-transparency.atlassian.net/browse/DEV-5839):
    - [X] Link to this Pull-Request
    - [ ] Performance evaluation of affected (API | Script | Download)
    - [ ] Before / After data comparison

**Area for explaining above N/A when needed:**
```
No changes to frontend, no changes to matviews, no ops tickets needed.
```